third_party/blink/web_tests/platform/generic/fast/css/invalidation/recalc-direct-adjacent-001-expected.txt

Check that we don't do unnecessary style recalcs for sibling trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(document.getElementById('green'), null).color is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
